Fix WIL project dependency
Dependencies are required to be declared somewhere in the directory
hierarchy that depends on them, so having a "dependencies" directory to
the side was a bad idea. The dependency issue was fixed by moving the
use of FetchContent for WIL from the side "dependencies" directory to
the src directory itself.

Transitive dependency on WIL was also achieved by listing it as a PUBLIC
lib dependency for the project.